+++ src/models/database_admin.rs
@@ -18,7 +18,7 @@ pub struct DatabaseAdmin {
     /// ID для каждого экземпляра пользователя базы данных. Автоматически
     /// генерируется при создании.
     #[serde(rename = "id")]
-    pub id:          f64,
+    pub id:          i64,
     /// Значение времени, указанное в комбинированном формате даты и времени
     /// ISO8601, которое представляет, когда была создана база данных.
     #[serde(rename = "created_at")]
@@ -42,7 +42,7 @@ pub struct DatabaseAdmin {
 impl DatabaseAdmin {
     /// Пользователь базы данных
     pub fn new(
-        id: f64,
+        id: i64,
         created_at: String,
         login: String,
         password: String,
